Update release doc with best practice for creating changelog (#1595)
# Description
## What is the current behavior?
Currently, we relied on milestones to form the changelogs but this can
lead us to miss important changes.

closes: #1593  

## What is the new behavior?
Going forward we can follow the below approach to avoid such
occurrences.

```
One way to identify what changed from the previous release is to use a comparison URL between the previous release branch and the main branch.

Example -
`https://github.com/astronomer/astro-sdk/compare/1.3.3...main`

This will show all the commits that are in the main branch but not in the release branch. Using this we can note all the changes significant enough to be mentioned in CHANGELOGS.

```

Add it as best practice. 

## Does this introduce a breaking change?
Nope


### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

Co-authored-by: Kaxil Naik <kaxilnaik@gmail.com>