Create unit tests for cleanup_table (#1587)
# Description
## What is the current behavior?
Currenlty the `test_cleanup` file tests take 16 minutes to run. This is
due to many external database calls and file loads/table creations. This
is causing us unnecessary expenses and slowing down development as we
wait for CI/CD to finish.


## What is the new behavior?
With this PR, we can use the unittest mocking library to mock out any
external SQL calls. Since we already know that the SQL calls will work,
we can focus purely on whether the astro library generates the correct
SQL and deletes the correct tables. This has caused a 50% reduction in
run-time (16 minutes to 8 minutes).

In the interest of not making the perfect the enemy of the good, I have
opted not to mock out the mapped tasks tests, as those seem more complex
and I believe there are lower hanging fruit we can address to reduce CI
runtime.

## Does this introduce a breaking change?
No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary