Release SQL CLI 0.3.2, setting Python SDK < 1.5 (#1703)
When installing the SQL CLI, we were installing any version of the
Python SDK that is newer than 1.3.1.

However, newer versions of the Python SDK may introduce breaking
changes. Additionally, there was a request in the past from Vikram for
us to bound specific versions of the SQL CLI with features of specific
versions of the Python SDK (e.g. Google Drive support).

This particular release is necessary, so users who install the newest
version of the Astro CLI & download the SQL CLI don't face issues
related to MSSQL dependencies introduced in the latest Python SDK
1.5 release