Add validation mode support in COPY INTO in snowflake (#1689)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
Currently, the load_file operator can hide errors when doing native
Snowflake transfers. We should raise an exception and display the errors
to the users, not hide this problem.

As it stands, a user may be loading 10k rows into a table, end up with
five rows, and not realise there is an issue.

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
closes: #581


## What is the new behaviour?
<!-- Please describe the behaviour or changes that are being added by
this PR. -->
- Add validation mode as part of the `COPY INTO` command. Specify the
supported validation mode; `RETURN_n_ROWS` or `RETURN_ERRORS` or
`RETURN_ALL_ERRORS`. This instructs the COPY command to validate the
data files instead of loading them into the specified table; i.e. the
COPY command tests the files for errors but does not load them. Read
more at:

https://docs.snowflake.com/en/sql-reference/sql/copy-into-table.html#optional-parameters
- Remove the `ON_ERROR=Continue` hardcoded for CSV file types
- Add an option for the user to pass `ON_ERROR=Continue` as part of
`copy_options`

Example:
```
aql.load_file(
        input_file=File("s3://astro-sdk/python_sdk/example_dags/data/sample.csv", conn_id="aws_conn"),
        output_table=Table(
            conn_id=SNOWFLAKE_CONN_ID,
        ),
        load_options=[
            SnowflakeLoadOptions(
                file_options={"SKIP_HEADER": 1, "SKIP_BLANK_LINES": True},
                copy_options={"ON_ERROR": "CONTINUE"},
                validation_mode="RETURN_ALL_ERRORS",
            )
        ],
    )
```

## Does this introduce a breaking change?
Yes

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README / documentation, if necessary