Use handler in run_sql (#1773)
**Please describe the feature you'd like to see**
Currently, the handler param in the run_sql() method is only used in
data bricks all other databases are not using it, which leads to
different response types since the handler is processed within
`run_sql()` for data bricks and not for other databases. Also, the
signature of `run_sql()` in data bricks deviates from the base. We need
to standardize and when we do, we can remove the below check as well.

**Describe the solution you'd like**
Ideally, we should have the same signatures for all the run_sql()
database implementations. This will prevent adding explicit conditions
for data bricks in other parts of the code that should ideally be
database agnostic.
example:
https://github.com/astronomer/astro-sdk/blob/2f8c8d0ccbff3cc13af78685cbefa39f473991c3/python-sdk/src/astro/sql/operators/raw_sql.py#L47

**Acceptance Criteria**
- [ ] Where ever we have used run_sql(), we shouldn't be having any
database-specific checks.

closes #1664