Running example dag on astro cloud with multiple workers (#1750)
**Please describe the feature you'd like to see**
Currently, we are using SQLite-based dag for transform, merge and append
when running on astro cloud since there can be more than 1 worker. When
a different worker is assigned to transform, merge or append then the
dag will fail with the table not found. and now we are setting the
worker slots to 1 so all tasks run on the same worker.
SQLite is not a network-based database so it doesn't have any network
connection ability built into it.
1.  Make SQLite access as mount volume to workers.
2. Find an existing library which does the point 1

**Describe the solution you'd like**
To increase the worker slot in astro cloud we need to change the SQLite
connection of these dags to other database connections.
Preferably Postgres.
Since duckdb also had the same issue we decided to do multiple
deployments for the sqlite and duckdb ones and others separately.
**Acceptance Criteria**

- [x] All checks and tests in the CI should pass
- [x] Run the dags on astro cloud for multiple workers
- [x] Unit tests (90% code coverage or more, [once
available](https://github.com/astronomer/astro-sdk/issues/191))
- [x] Integration tests (if the feature relates to a new database or
external service)
- [x] Example DAG
- [x] Docstrings in
[reStructuredText](https://peps.python.org/pep-0287/) for each of
methods, classes, functions and module-level attributes (including
Example DAG on how it should be used)
- [x] Exception handling in case of errors
- [x] Logging (are we exposing useful information to the user? e.g.
source and destination)
- [x] Improve the documentation (README, Sphinx, and any other relevant)

- [x] How to use Guide for the feature
([example](https://airflow.apache.org/docs/apache-airflow-providers-postgres/stable/operators/postgres_operator_howto_guide.html))

closes #1737