Check if connection is minio then set use_native_support=False (#1767)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
When the user passes the s3 url but has minio as the setup, for minio to
any database transfers by default it uses the native path and fails as
there is no native path. There is no direct way to check if the
connection is of type minio. Introducing a parameter `minio` in extras
as part of connections would help to recognize the connection type

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
closes: #1746

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Check if the AWS connection has `minio` in `extra` then set
use_native_support=False
- Example minio conn:
```
  - conn_id: minio_conn
    conn_type: aws
    description: null
    extra:
      aws_access_key_id: "dummy"
      aws_secret_access_key: "dummy"
      endpoint_url: "http://127.0.0.1:9000"
      minio: True
```


## Does this introduce a breaking change?
No

### Checklist
- [x] Created tests which fail without the change (if possible)
- [x] Extended the README / documentation, if necessary

---------

Co-authored-by: utkarsh sharma <utkarsharma2@gmail.com>