Fix astro flow deploy integration by having a conditional sql_cli import
The astro flow deploy command fails currently as it cannot find
the sql_cli library installed. This library is required only for
local runs when running a single task. So, to fix the deploy command,
we identify if it for a local run using the `task_to_execute` list
in the context which is populated only for single task runs. This
list will be empty when deploy DAGs that are supposed to generate
and run wholly.