Fix astro flow deploy integration by having a conditional sql_cli import (#1841)
# Description
## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
The `astro flow deploy` command fails currently as it cannot find the
sql_cli library installed

<!--
Issues are required for both bug fixes and features.
Reference it using one of the following:

closes: #ISSUE
related: #ISSUE
-->
closes: #1825 

## What is the new behavior?
This library `sql_cli` is required only for local runs when running a
single task. So, to fix the deploy command, we identify if it is for a
local run using the `task_to_execute` list in the context populated only
for single task runs. This list will be empty when deploying DAGs that
are supposed to generate and run wholly.

## Does this introduce a breaking change?
No.

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary