Fix failing test - test_aql_load_file_local_file_pattern_dataframe on CI (#1895)
# Description
## What is the current behavior?
Currently, the test case is failing on the main branch -
`test_aql_load_file_local_file_pattern_dataframe`.

## What is the new behavior?
A test case is failing because the index was considered in the equality
test when using `sort_value()`. With this change we are ignoring the
index. Since we are concatenating two files, if the files are
concatenated in the right order we have a test passing if they don't it
fails.

## Does this introduce a breaking change?
Nope

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary