Add a Session class that will allow users to create pre-queries and post-queries (#1898)
# Description
## What is the current behavior?
Currently there is no way to include "pre-queries" and "post-queries" to
queries generated by transform and raw_sql function. This issue is
notably an issue in https://github.com/astronomer/astro-sdk/issues/1820
where the user needs to create a new connection every time they want to
add a query tag to a query. This creates a lot of clutter as a single
team can need 10 unique connections dependent on the query tags needed.

closes: https://github.com/astronomer/astro-sdk/issues/1820



## What is the new behavior?
The new behavior is to open up "pre_queries" and "post_queries" to the
user. This will allow users to generalize a single "Session" object that
can be passed to multiple functions without cluttering the returned SQL
statement. This session object can also be modified based on Airflow
variables.

## Does this introduce a breaking change?

No

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary