Allow users to disable schema check and creation on `load_file` (#1922)
Support running `load_file` without checking if the table schema exists
or trying to create it.

Recently a user reported that the cost of checking if the schema exists
is very high for Snowflake:
"I have a (`load_file`) task that took 1:36 minutes to run, and it was
1:30 running the information schema query."
This is likely happening for other databases as well.

Introduce two ways of disabling schema checks:

1. On a per-task basis, by exposing the argument `schema_exists` in
`aql.load_file`
When this argument is `True`, the SDK will not check if the schema
exists or try to create it.
It is `False` by default, and the Python SDK will behave as of 1.6
(running schema check and, if needed, trying to create the schema)

2. Globally, by exposing the Airflow configuration
`load_table_schema_exists` in the `[astro-sdk]` section. This can also
be set using the environment variable
`AIRFLOW__ASTRO_SDK__LOAD_TABLE_SCHEMA_EXISTS`. The global configuration
can be overridden per task, using [1].

Closes: #1921