Refactor `assume_schema_exists` and expose it in `aql.transform` (#1925)
This is a follow-up for #1922. In that PR we allowed users to skip
schema check & creation for `aql.load_file`, but we missed the fact that
`aql.transform` and `aql.transform_file` had the same issue. This PR
aims to address this limitation.

Changes included in this PR:
* Rename config `load_table_schema_exists` to `assume_schema_exists`
* Rename (`load_file`) argument `schema_exists` to
`assume_schema_exists`
* Refactor where the check for `assume_schema_exists` happens. Before,
it happened only inside the `load_file_to_table`. Now, it is part of
`create_schema_if_applicable`. This makes this feature available in the
`aql.transform` task as well
* Rename `Database.create_schema_if_needed` to
`Database.create_schema_if_applicable`
* Expose `assume_schema_exists` in `aql.transform`
* Release 1.7.0a2