Set load_secret True in in Databricks S3 load file DAG (#2071)
closes: https://github.com/astronomer/astro-sdk/issues/2070
closes: https://github.com/astronomer/astro-sdk/issues/2068

The default value for load_secret is false 
and because of that Databricks cluster was 
not able to access the AWS S3 file. 
I have changed that to True in this PR and 
the Airflow task is green now.
but not sure if this is because of some recent changes in Databricks.