Deploy only if RC is  released (#2074)
Our RC test workflow was updating the deployment and running the master
DAG even if there was no RC release. This PR only initiates deployment
and DAG triggers when we have an RC release. Additionally, I have
updated the cron of the deploy job to avoid both workflows running at
the same time.